requirements.txt
langchain==1.2.10
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-openai==1.1.9
langchain-text-splitters==1.1.1
python-dotenv==1.2.1
chromadb==1.5.0
openai==2.21.0
tiktoken==0.12.0
streamlit==1.54.0
pypdf==6.7.0
requests==2.32.5
Pillow==12.1.1
psycopg[binary]==3.2.10
